refactor: split out context via object literal
