+++ src/context/autoFill.js
@@ -1,0 +1,32 @@
+export default {
+  autoFill: (state, action) => {
+    return {
+      ...state,
+      contentBlocks: state.contentBlocks.map((item, index) => {
+        if (index !== action.index) return item
+        return {
+          ...item,
+          content: {
+            ...action.payload
+          }
+        }
+      })
+    }
+  },
+  autoFillInfoStrip: (state, action) => {
+    let name = action.type.slice(8)
+    let blockType = name.charAt(0).toLowerCase() + name.slice(1)
+    return {
+      ...state,
+      [blockType]: [...action.payload]
+    }
+  },
+  autoFillCategories: (state, action) => {
+    let name = action.type.slice(8)
+    let blockType = name.charAt(0).toLowerCase() + name.slice(1)
+    return {
+      ...state,
+      [blockType]: [...action.payload]
+    }
+  }
+}
