+++ src/context/saveFile.js
@@ -1,0 +1,8 @@
+import saveFileFunc from '../lib/saveFile'
+
+export default {
+  saveFile: state => {
+    saveFileFunc(state)
+    return { ...state }
+  }
+}
